pip-requirements.txt
pytz
python-memcached>=1.3 
Django>=4.1.2,<4.2.0
mysqlclient==2.1.1
django-crispy-forms==1.14.0
crispy-bootstrap5==0.6
tablib==3.2.1
django-filter==22.1
django-tables2==2.4.1
django-bootstrap-v5==1.0.11
django-import-export==3.0.2
django-auto-logout==0.5.0
django-select2==8.0.0
django-redis==5.2.0